core-sparta/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
openai>=0.27.0
pydantic-settings>=0.3.0
python-dotenv>=1.0.0
